Moved RenderRouteWithOutletContext to src/test
